+++ src/applications/edu-benefits/10297/config/prefill-transformer.js
@@ -73,7 +73,9 @@ export default function prefillTransformer(pages, formData, metadata, state) {
 
   return {
     metadata,
-    formData: finalFormData,
+    formData: {
+      ...finalFormData,
+    },
     pages,
   };
 }
